+++ changelog.md
@@ -4,6 +4,7 @@
 
 - **New:**
   - Character manager: Added `Disconnect on load` for each character allowing you to force the current connection to disconnect and using the new characters auto connect settings
+  - Code Editor: add armour of holding support to area designer and code template
 - **Changed:**
   - Update electron 13.1.1 to 13.1.2